webapp: refactor
